refactor: move package of ApplicantInformation
enter the commit message for your changes. Lines starting
